+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## [0.1.6]
+- Show 'Azure charge' hint on disconnect from azurecontainer.io based Spots
+- Support setting the file watch path on the Spot Host
+
 ## [0.1.5]
 - Initial public release